refactor: 통합 테스트 설정 변경
